+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/process/test/ScenarioTestService.scala
@@ -1,5 +1,6 @@
 package pl.touk.nussknacker.ui.process.test
 
+import cats.syntax.either._
 import com.carrotsearch.sizeof.RamUsageEstimator
 import com.typesafe.scalalogging.LazyLogging
 import pl.touk.nussknacker.engine.api.definition.{DualParameterEditor, Parameter, StringParameterEditor}
@@ -19,8 +20,7 @@ import pl.touk.nussknacker.ui.api.description.NodesApiEndpoints.Dtos.TestSourceP
 import pl.touk.nussknacker.ui.definition.DefinitionsService
 import pl.touk.nussknacker.ui.process.deployment.ScenarioTestExecutorService
 import pl.touk.nussknacker.ui.process.marshall.CanonicalProcessConverter
-import pl.touk.nussknacker.ui.process.test.ScenarioTestService.SourceTestError
-import pl.touk.nussknacker.ui.process.test.ScenarioTestService.SourceTestError.TooManySamplesRequestedError
+import pl.touk.nussknacker.ui.process.test.ScenarioTestService.{GenerateTestDataError, SourceTestError}
 import pl.touk.nussknacker.ui.processreport.{NodeCount, ProcessCounter, RawCount}
 import pl.touk.nussknacker.ui.security.api.LoggedUser
 import pl.touk.nussknacker.ui.uiresolving.UIProcessResolver
@@ -75,15 +75,17 @@ class ScenarioTestService(
       testSampleSize: Int
   )(
       implicit user: LoggedUser
-  ): Either[String, RawScenarioTestData] = {
+  ): Either[GenerateTestDataError, RawScenarioTestData] = {
     val canonical = toCanonicalProcess(scenarioGraph, processVersion, isFragment)
 
     for {
-      _ <- validateSampleSize(testSampleSize).left.map(error =>
-        s"Too many samples requested, limit is ${error.maxSamples}"
-      )
-      generatedData <- testInfoProvider.generateTestData(processVersion, canonical, testSampleSize).left.map(_.message)
-      rawTestData   <- preliminaryScenarioTestDataSerDe.serialize(generatedData)
+      _ <- validateSampleSize(testSampleSize)(GenerateTestDataError.TooManySamplesRequestedError)
+      generatedData <- testInfoProvider
+        .generateTestData(processVersion, canonical, testSampleSize)
+        .leftMap(GenerateTestDataError.ScenarioTestDataGenerationError)
+      rawTestData <- preliminaryScenarioTestDataSerDe
+        .serialize(generatedData)
+        .leftMap(GenerateTestDataError.ScenarioTestDataSerializationError)
     } yield rawTestData
   }
 
@@ -93,11 +95,10 @@ class ScenarioTestService(
       size: Int
   ): Either[SourceTestError, RawScenarioTestData] = {
     for {
-      _ <- validateSampleSize(size)
+      _ <- validateSampleSize(size)(SourceTestError.TooManySamplesRequestedError)
       result <- testInfoProvider
         .generateTestDataForSource(metaData, sourceNodeData, size)
-        .left
-        .map {
+        .leftMap {
           case SourceTestDataGenerationError.SourceCompilationError(nodeId, errors) =>
             SourceTestError.SourceCompilationError(nodeId, errors.map(_.toString))
           case SourceTestDataGenerationError.UnsupportedSourceError(nodeId) =>
@@ -107,8 +108,7 @@ class ScenarioTestService(
         }
       rawTestData <- preliminaryScenarioTestDataSerDe
         .serialize(result)
-        .left
-        .map(msg => SourceTestError.SerializationError(s"Failed to serialize test data: $msg"))
+        .leftMap(serializationError => SourceTestError.ScenarioTestDataSerializationError(serializationError))
     } yield rawTestData
   }
 
@@ -121,7 +121,8 @@ class ScenarioTestService(
     for {
       preliminaryScenarioTestData <- preliminaryScenarioTestDataSerDe
         .deserialize(rawTestData)
-        .fold(error => Future.failed(new IllegalArgumentException(error)), Future.successful)
+        // TODO ljd
+        .fold(error => Future.failed(new IllegalArgumentException(error.toString)), Future.successful)
       canonical = toCanonicalProcess(
         scenarioGraph,
         processVersion,
@@ -136,9 +137,7 @@ class ScenarioTestService(
         canonical,
         scenarioTestData,
       )
-      _ <- {
-        assertTestResultsAreNotTooBig(testResults)
-      }
+      _ <- assertTestResultsAreNotTooBig(testResults)
     } yield ResultsWithCounts(testResults, computeCounts(canonical, isFragment, testResults))
   }
 
@@ -159,11 +158,11 @@ class ScenarioTestService(
     } yield ResultsWithCounts(testResults, computeCounts(canonical, isFragment, testResults))
   }
 
-  private def validateSampleSize(size: Int): Either[TooManySamplesRequestedError, Unit] = {
+  private def validateSampleSize[E](size: Int)(tooManySamplesError: Int => E): Either[E, Unit] = {
     Either.cond(
       size <= testDataSettings.maxSamplesCount,
       (),
-      SourceTestError.TooManySamplesRequestedError(testDataSettings.maxSamplesCount)
+      tooManySamplesError(testDataSettings.maxSamplesCount)
     )
   }
 
@@ -195,7 +194,7 @@ class ScenarioTestService(
     val testDataResultApproxByteSize = RamUsageEstimator.sizeOf(testResults)
     if (testDataResultApproxByteSize > testDataSettings.resultsMaxBytes) {
       logger.info(
-        s"Test data limit exceeded. Approximate test data size: $testDataResultApproxByteSize, but limit is: ${testDataSettings.resultsMaxBytes}"
+        s"Test results limit exceeded. Approximate size: $testDataResultApproxByteSize, but limit is: ${testDataSettings.resultsMaxBytes}"
       )
       Future.failed(new RuntimeException("Too much test data. Please decrease test input data size."))
     } else {
@@ -218,14 +217,25 @@ class ScenarioTestService(
 }
 
 object ScenarioTestService {
+  sealed trait GenerateTestDataError
+
+  object GenerateTestDataError {
+    final case class ScenarioTestDataGenerationError(cause: TestInfoProvider.ScenarioTestDataGenerationError)
+        extends GenerateTestDataError
+    final case class ScenarioTestDataSerializationError(cause: PreliminaryScenarioTestDataSerDe.SerializationError)
+        extends GenerateTestDataError
+    final case class TooManySamplesRequestedError(maxSamples: Int) extends GenerateTestDataError
+  }
+
   sealed trait SourceTestError
 
   object SourceTestError {
     final case class SourceCompilationError(nodeId: String, errors: List[String]) extends SourceTestError
     final case class UnsupportedSourcePreviewError(nodeId: String)                extends SourceTestError
     case object NoDataGeneratedError                                              extends SourceTestError
-    final case class SerializationError(message: String)                          extends SourceTestError
-    final case class TooManySamplesRequestedError(maxSamples: Int)                extends SourceTestError
+    final case class ScenarioTestDataSerializationError(cause: PreliminaryScenarioTestDataSerDe.SerializationError)
+        extends SourceTestError
+    final case class TooManySamplesRequestedError(maxSamples: Int) extends SourceTestError
   }
 
 }